feat: add method GET, PUT, and DELETE

+++ src/model/lesson.rs
@@ -1,5 +1,4 @@
-// src/model/Lesson.rs
-
+// src/model/lesson.rs
 use serde::{Serialize, Deserialize};
 use mongodb::bson::oid::ObjectId;
 
@@ -8,6 +7,6 @@ pub struct Lesson {
     #[serde(rename = "_id", skip_serializing_if = "Option::is_none")]
     pub id: Option<ObjectId>,
     pub name: String,
-    pub credits: u8,
+    pub credits: u32, 
     pub teacher: String,
 }
